move store.ts in src directory
