+++ src/index.tsx
@@ -2,8 +2,8 @@ import React from "react";
 import ReactDOM from "react-dom";
 import { Provider } from "react-redux";
 import { Davit } from "./pages/Davit";
-import { store } from "./app/store";
 import * as serviceWorker from "./serviceWorker";
+import { store } from "./store";
 import "./style/index.css";
 
 ReactDOM.render(
