+++ src/pages/Davit.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from "@testing-library/react";
 import React from "react";
 import { Provider } from "react-redux";
-import { store } from "../app/store";
+import { store } from "../store";
 import { Davit } from "./Davit";
 
 test("Davit basic UI test", () => {
