+++ src/slices/GlobalSlice.ts
@@ -1,7 +1,7 @@
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
-import { AppThunk, RootState } from "../app/store";
 import { DataAccess } from "../dataAccess/DataAccess";
 import { DataAccessResponse } from "../dataAccess/DataAccessResponse";
+import { AppThunk, RootState } from "../store";
 
 interface GlobalState {
     errors: string[];
