+++ src/slices/thunks/ActionThunks.ts
@@ -1,8 +1,8 @@
-import { AppThunk } from "../../app/store";
 import { SequenceStepCTO } from "../../dataAccess/access/cto/SequenceStepCTO";
 import { ActionTO } from "../../dataAccess/access/to/ActionTO";
 import { DataAccess } from "../../dataAccess/DataAccess";
 import { DataAccessResponse } from "../../dataAccess/DataAccessResponse";
+import { AppThunk } from "../../store";
 import { editActions, EditActions, Mode } from "../EditSlice";
 import { GlobalActions } from "../GlobalSlice";
 import { MasterDataActions } from "../MasterDataSlice";
