+++ src/slices/thunks/ActorThunks.ts
@@ -1,7 +1,7 @@
-import { AppThunk } from "../../app/store";
 import { ActorCTO } from "../../dataAccess/access/cto/ActorCTO";
 import { DataAccess } from "../../dataAccess/DataAccess";
 import { DataAccessResponse } from "../../dataAccess/DataAccessResponse";
+import { AppThunk } from "../../store";
 import { editActions, Mode } from "../EditSlice";
 import { GlobalActions } from "../GlobalSlice";
 import { MasterDataActions } from "../MasterDataSlice";
