+++ src/slices/thunks/ChainLinkThunks.ts
@@ -1,9 +1,9 @@
-import { AppThunk } from "../../app/store";
 import { ChainDecisionTO } from "../../dataAccess/access/to/ChainDecisionTO";
 import { ChainlinkTO } from "../../dataAccess/access/to/ChainlinkTO";
 import { GoToTypesChain } from "../../dataAccess/access/types/GoToTypeChain";
 import { DataAccess } from "../../dataAccess/DataAccess";
 import { DataAccessResponse } from "../../dataAccess/DataAccessResponse";
+import { AppThunk } from "../../store";
 import { editActions, Mode } from "../EditSlice";
 import { GlobalActions } from "../GlobalSlice";
 import { MasterDataActions } from "../MasterDataSlice";
