+++ src/slices/thunks/ConditionThunks.ts
@@ -1,8 +1,8 @@
 // ----------------------------------------------- DECISION -----------------------------------------------
 
-import { AppThunk } from "../../app/store";
 import { ConditionTO } from "../../dataAccess/access/to/ConditionTO";
 import { DecisionTO } from "../../dataAccess/access/to/DecisionTO";
+import { AppThunk } from "../../store";
 import { DavitUtil } from "../../utils/DavitUtil";
 import { editActions } from "../EditSlice";
 import { EditDecision } from "./DecisionThunks";
