+++ src/slices/thunks/DataSetupThunks.ts
@@ -1,7 +1,7 @@
-import { AppThunk } from "../../app/store";
 import { DataSetupCTO } from "../../dataAccess/access/cto/DataSetupCTO";
 import { DataAccess } from "../../dataAccess/DataAccess";
 import { DataAccessResponse } from "../../dataAccess/DataAccessResponse";
+import { AppThunk } from "../../store";
 import { editActions, Mode } from "../EditSlice";
 import { GlobalActions } from "../GlobalSlice";
 import { MasterDataActions } from "../MasterDataSlice";
