+++ src/slices/thunks/GroupThunks.ts
@@ -1,7 +1,7 @@
-import { AppThunk } from "../../app/store";
 import { GroupTO } from "../../dataAccess/access/to/GroupTO";
 import { DataAccess } from "../../dataAccess/DataAccess";
 import { DataAccessResponse } from "../../dataAccess/DataAccessResponse";
+import { AppThunk } from "../../store";
 import { editActions, Mode } from "../EditSlice";
 import { GlobalActions } from "../GlobalSlice";
 import { MasterDataActions } from "../MasterDataSlice";
