+++ src/slices/thunks/StepThunks.ts
@@ -1,10 +1,10 @@
-import { AppThunk } from "../../app/store";
 import { SequenceCTO } from "../../dataAccess/access/cto/SequenceCTO";
 import { SequenceStepCTO } from "../../dataAccess/access/cto/SequenceStepCTO";
 import { DecisionTO } from "../../dataAccess/access/to/DecisionTO";
 import { GoToTypes } from "../../dataAccess/access/types/GoToType";
 import { DataAccess } from "../../dataAccess/DataAccess";
 import { DataAccessResponse } from "../../dataAccess/DataAccessResponse";
+import { AppThunk } from "../../store";
 import { DavitUtil } from "../../utils/DavitUtil";
 import { editActions, Mode } from "../EditSlice";
 import { GlobalActions } from "../GlobalSlice";
